Break into multiple files, import
Also detach ChooseFonts from the Updater because it is completely unrelated

quick fix
